refactor: transforma hook usePagarme em um serviço
