+++ .eslintrc.json
@@ -56,7 +56,7 @@
       "no-use-before-define": ["error", { "functions": false }],
       "global-require": "off"
     },
-    "ignorePatterns": ["public/", "node_modules/", ".cache/", "./functions/", "static/"],
+    "ignorePatterns": ["public/", "node_modules/", ".cache/", "./functions/", "./__mocks__/", "static/"],
     "settings": {
       "react": {
         "version": "detect"
